# openapi.yml
...
openapi: 3.1.0
info:
  title: CubeAPI
  description: E2B-compatible sandbox API server.
  license:
    name: ''
  version: 0.1.0
servers:
- url: /cubeapi/v1
  description: CubeAPI SDK/dashboard surface (source of truth generated via `cube-api --export-openapi`)
paths:
...
  /sandboxes:
    get:
      tags:
      - handlers::sandboxes
      operationId: list_sandboxes
      parameters:
      - name: metadata
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Sandbox list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListedSandbox'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - handlers::sandboxes
      operationId: create_sandbox
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewSandbox'
        required: true
      responses:
        '201':
          description: Sandbox created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Template or snapshot not found (e.g. deleted/tombstoned)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: A referenced resource is in a conflicting state and cannot be used to create a sandbox
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /sandboxes/{sandboxID}/logs:
    get:
      tags:
      - handlers::sandboxes
      operationId: get_sandbox_logs
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      - name: start
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Sandbox logs (legacy shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxLogs'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v2/sandboxes:
    get:
      tags:
      - handlers::sandboxes
      operationId: list_sandboxes_v2
      parameters:
      - name: metadata
        in: query
        required: false
        schema:
          type: string
      - name: state
        in: query
        required: false
        schema:
          type: string
      - name: nextToken
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Sandbox list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListedSandbox'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v2/sandboxes/{sandboxID}/logs:
    get:
      tags:
      - handlers::sandboxes
      operationId: get_sandbox_logs_v2
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      - name: cursor
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: direction
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Structured sandbox logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxLogsV2Response'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    E2BNetworkRule:
      type: object
      properties:
        transform:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/E2BNetworkTransform'
      additionalProperties: false
    E2BNetworkTransform:
      type: object
      properties:
        headers:
          type: object
          additionalProperties:
            type: string
          propertyNames:
            type: string
      additionalProperties: false
    EgressRule:
      type: object
      description: 'L7 egress rule: match conditions + action (allow/deny, audit, credential injection).'
      required:
      - name
      - match
      - action
      properties:
        action:
          $ref: '#/components/schemas/EgressRuleAction'
        match:
          $ref: '#/components/schemas/EgressRuleMatch'
        name:
          type: string
          description: Human-readable label used for audit logging.
    EgressRuleAction:
      type: object
      description: |-
        Rule action.

        - `allow=true`: pass the request through; optional credential injection.
        - `allow=false`: reject (HTTP 403); `inject` is ignored if set.
        - `audit` defaults to `"metadata"` server-side when omitted.
      required:
      - allow
      properties:
        allow:
          type: boolean
        audit:
          type:
          - string
          - 'null'
          description: One of `"full"`, `"metadata"`, `"none"`.
        inject:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/EgressRuleInject'
    EgressRuleInject:
      type: object
      description: |-
        Credential injection. Only honored when `Action.allow=true` and the
        request is HTTPS with matching SNI/Host (downstream enforces).
      required:
      - header
      - secret
      properties:
        format:
          type:
          - string
          - 'null'
          description: Template containing `${SECRET}`; defaults to `"${SECRET}"` when omitted.
        header:
          type: string
        secret:
          type: string
    EgressRuleMatch:
      type: object
      description: |-
        Rule match conditions. All fields optional; empty match matches any request.

        Multi-field semantics: AND across fields, OR within `method`.
        Comparisons on sni/host/scheme are case-insensitive.

        `port` + `scheme` together pin the (host, port) tuple CubeEgress intercepts.
        Both nil keeps the legacy default {80/http, 443/https}. When `port` is set,
        `scheme` MUST also be set — same-`(host, port)` rules across the policy
        must agree on `scheme` (the server rejects the whole policy on mismatch).
      properties:
        host:
          type:
          - string
          - 'null'
        method:
          type:
          - array
          - 'null'
          items:
            type: string
        path:
          type:
          - string
          - 'null'
        port:
          type:
          - integer
          - 'null'
          format: int32
        scheme:
          type:
          - string
          - 'null'
        sni:
          type:
          - string
          - 'null'
    HashMap:
      type: object
      additionalProperties:
        type: string
      propertyNames:
        type: string
...
    ListedSandbox:
      type: object
      description: One entry in GET /sandboxes (RunningSandbox in OpenAPI spec).
      required:
      - templateID
      - sandboxID
      - clientID
      - startedAt
      - cpuCount
      - memoryMB
      - state
      - envdVersion
      properties:
        alias:
          type:
          - string
          - 'null'
        clientID:
          type: string
        cpuCount:
          type: integer
          format: int32
          description: Legacy whole-vCPU count. Use cpuMilli for fractional CPU values.
        cpuMilli:
          type: integer
          format: int32
          description: Exact CPU allocation in millicores, e.g. 100 or 500.
        diskSizeMB:
          type:
          - integer
          - 'null'
          format: int32
        endAt:
          type:
          - string
          - 'null'
          format: date-time
          description: |-
            Projected next-timeout instant. Omitted for never-timeout sandboxes
            (no deadline) rather than being misreported as equal to startedAt.
        envdVersion:
          type: string
        memoryMB:
          type: integer
          format: int32
          description: Memory in MiB. The MB suffix is retained for API compatibility.
        metadata:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/HashMap'
        sandboxID:
          type: string
        startedAt:
          type: string
          format: date-time
        state:
          $ref: '#/components/schemas/SandboxState'
        templateID:
          type: string
        volumeMounts:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/SandboxVolumeMount'
    LogLevel:
      type: string
      enum:
      - debug
      - info
      - warn
      - error
    NewSandbox:
      type: object
      description: |-
        Request body for POST /sandboxes
        Field names match exactly what the E2B SDK sends.
        Rule: ID abbreviations → uppercase (templateID, sandboxID, envVars);
              allow_internet_access is a known SDK snake_case quirk;
              lifecycle is a nested object — see SandboxLifecycleConfig.
        `envVars` is the canonical field name; `envs` is accepted as a compatibility
        alias for E2B SDK callers.
      required:
      - templateID
      properties:
        allow_internet_access:
          type:
          - boolean
          - 'null'
          description: SDK sends this as snake_case (known quirk).
        autoPause:
          type:
          - boolean
          - 'null'
          description: |-
            e2b SDK compatibility: the SDK flattens its user-facing `lifecycle`
            object into top-level `autoPause` / `autoResume` before it hits the
            wire, so `lifecycle` never arrives from an SDK caller. Ignored when
            `lifecycle` is present.
        autoResume:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/SandboxAutoResume'
            description: See `auto_pause`. Only meaningful alongside `autoPause = true`.
        distributionScope:
          type:
          - array
          - 'null'
          items:
            type: string
        envVars:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/HashMap'
        lifecycle:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/SandboxLifecycleConfig'
            description: |-
              Sandbox lifecycle configuration. Maps to e2b's `lifecycle` object so
              callers that already speak e2b can pass through unchanged. Absent
              (None) means today's behaviour: idle sandboxes are killed.
        mcp: {}
        metadata:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/HashMap'
        network:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/SandboxNetworkConfig'
        secure:
          type:
          - boolean
          - 'null'
        templateID:
          type: string
        timeout:
          type:
          - integer
          - 'null'
          format: int32
          description: Optional idle TTL in seconds. See docs/guide/lifecycle.md.
        volumeMounts:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/SandboxVolumeMount'
...
    Sandbox:
      type: object
      description: |-
        Response for POST /sandboxes and POST /sandboxes/{id}/connect.
        All ID abbreviations uppercase per E2B OpenAPI spec.
      required:
      - templateID
      - sandboxID
      - clientID
      - envdVersion
      properties:
        alias:
          type:
          - string
          - 'null'
        clientID:
          type: string
        domain:
          type:
          - string
          - 'null'
        envdAccessToken:
          type:
          - string
          - 'null'
        envdVersion:
          type: string
        sandboxID:
          type: string
        templateID:
          type: string
        trafficAccessToken:
          type:
          - string
          - 'null'
    SandboxAutoResume:
      oneOf:
      - type: boolean
      - type: object
        required:
        - enabled
        properties:
          enabled:
            type: boolean
      description: |-
        Wire shape of the top-level `autoResume` field the e2b SDK sends. Current
        SDK releases send `{"enabled": true}`; older ones and hand-rolled clients
        send a bare `true`. Both mean the same thing, so accept either rather than
        rejecting on shape.
...
    SandboxLifecycleConfig:
      type: object
      description: |-
        Sandbox lifecycle configuration. Mirrors the e2b SDK's `lifecycle` object —
        see https://e2b.dev/docs/sandbox/auto-resume for the canonical reference.

        `on_timeout` decides what happens when the sandbox idle timer fires; the
        historical default is "kill" (delete the sandbox) which matches today's
        behaviour. `auto_resume` only takes effect when `on_timeout = "pause"` —
        it tells the proxy/sidecar to wake a paused sandbox up automatically when
        activity arrives, instead of returning an error.
      properties:
        autoResume:
          type: boolean
          description: |-
            Auto-resume on activity. Defaults to false. Only meaningful when
            `on_timeout` is set to "pause".
        onTimeout:
          $ref: '#/components/schemas/SandboxOnTimeout'
          description: '"kill" (default) | "pause".'
    SandboxLog:
      type: object
      description: Single raw log line — matches E2B SandboxLog schema (timestamp + line).
      required:
      - timestamp
      - line
      properties:
        line:
          type: string
        timestamp:
          type: string
          format: date-time
    SandboxLogEntry:
      type: object
      description: Structured log entry (v2 logs).
      required:
      - timestamp
      - message
      - level
      - fields
      properties:
        fields:
          type: object
          additionalProperties:
            type: string
          propertyNames:
            type: string
        level:
          $ref: '#/components/schemas/LogLevel'
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    SandboxLogs:
      type: object
      description: Legacy log response — matches E2B SandboxLogs schema.
      required:
      - logs
      - logEntries
      properties:
        logEntries:
          type: array
          items:
            $ref: '#/components/schemas/SandboxLogEntry'
        logs:
          type: array
          items:
            $ref: '#/components/schemas/SandboxLog'
    SandboxLogsV2Response:
      type: object
      description: v2 log response.
      required:
      - logs
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/SandboxLogEntry'
    SandboxNetworkConfig:
      type: object
      description: Network configuration for sandbox egress/ingress control.
      properties:
        allowOut:
          type:
          - array
          - 'null'
          items:
            type: string
        allowPublicTraffic:
          type:
          - boolean
          - 'null'
        denyOut:
          type:
          - array
          - 'null'
          items:
            type: string
        maskRequestHost:
          type:
          - string
          - 'null'
          description: |-
            Host authority forwarded to user services by CubeProxy. `${PORT}` is
            expanded to the requested sandbox port; envd traffic is exempt.
        rules:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/SandboxNetworkRulesInput'
            description: |-
              L7 egress rules. Accepts CubeSandbox's ordered rule array and E2B's
              host-keyed transform map for compatibility. Both are normalized to the
              ordered internal rule model.
    SandboxNetworkRulesInput:
      oneOf:
      - type: array
        items:
          $ref: '#/components/schemas/EgressRule'
      - type: object
        additionalProperties:
          type: array
          items:
            $ref: '#/components/schemas/E2BNetworkRule'
        propertyNames:
          type: string
      description: |-
        Accepted wire shapes for `SandboxNetworkConfig.rules`.

        CubeSandbox clients use the ordered rule array. E2B clients use a
        host-keyed map whose values are request transforms. CubeAPI normalizes both
        shapes into the internal ordered egress-rule model.
    SandboxOnTimeout:
      type: string
      enum:
      - kill
      - pause
    SandboxState:
      type: string
      description: State of the sandbox (running | paused | pausing | unknown)
      enum:
      - running
      - paused
      - pausing
      - unknown
    SandboxVolumeMount:
      type: object
      description: Volume mount inside the sandbox.
      required:
      - name
      - path
      properties:
        name:
          type: string
        path:
          type: string
        readOnly:
          type: boolean
          description: |-
            CubeSandbox extension: mount this volume read-only for this sandbox
            attachment. Defaults to false when omitted.
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...